Untangle gov and non-gov DRR summary panels. Mainly move the shared summary component to its own file and isolate the text keys.
