+++ client/src/panels/panel_declarations/results/CommonDrrSummary.scss
