--- client/src/panels/panel_declarations/results/drr_summary_text.js
@@ -1,5 +1,0 @@
-import { create_text_maker_component } from "src/components/index";
-
-import text from "./drr_summary_text.yaml";
-
-export const { text_maker, TM } = create_text_maker_component(text);
